+++ src/main/java/com/wormhole_xtreme/wormhole/model/StargateAnimator.java
@@ -602,7 +602,7 @@ class StargateAnimator
         }
         GateSounds.locked(gate);
         WormholeXTreme.getScheduler().scheduleSyncDelayedTask(WormholeXTreme.getThisPlugin(),
-            new StargateUpdateRunnable(gate, ActionToTake.ANIMATE_WOOSH));
+            new StargateUpdateRunnable(gate, ActionToTake.ANIMATE_WOOSH), 1L);
     }
 
     /**